feat: adding wallet widget and update crypto-list branch
