+++ lib/core/provider.dart
@@ -1,0 +1,5 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+final viewWalletValueProvider = StateProvider<bool>(
+  (ref) => false,
+);
