+++ lib/main.dart
@@ -1,9 +1,7 @@
-import 'dart:js';
-
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:warren_task_one/home_page/home_page.dart';
+import 'home_page/home_page.dart';
 
 void main() {
   runApp(DevicePreview(builder: (context) {
@@ -21,7 +19,7 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'Flutter Demo',
       theme: ThemeData(),
-      home: const WalletPage(),
+      home: const HomePage(),
     ));
   }
 }
